requirements.txt
# Dependencies for CLI Chat Tool
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
prompt-toolkit==3.0.39
rich==13.5.2
typer==0.9.0
pydantic==2.4.2
google-generativeai==0.8.5